refactor provider icons
